requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
commonmark==0.9.1
discord.py-self==2.0.0a1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pygments==2.11.2
rich==12.0.0
typing_extensions==4.1.1
yarl==1.7.2
